requirements.txt
fastapi
uvicorn
torch
torchvision
pillow
langgraph
langchain
langchain-openai
pydantic
python-multipart
streamlit
tqdm
matplotlib
seaborn
scikit-learn